+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * #3764: validate the topic refenced by a subscription actually exists
 * #3868: Add support for specifying maxConsumers for a subscription
 * #3873: Require TLSv1.2 or newer for external TLS endpoints
+* #3867: Add addition checks for router mesh status
 
 ## 0.30.2
 *  #2714: Allow setting security context of pods using persistent volumes